+++ app/src/main/java/com/example/group28assignmentapp/CommonUtils.java
@@ -4,7 +4,6 @@ import android.app.Dialog;
 import android.content.Context;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
-import android.graphics.drawable.Drawable;
 
 public class CommonUtils {
     public static Dialog showLoadingDialog(Context context) {
