+++ app/src/main/java/com/example/group28assignmentapp/MainActivity.java
@@ -1,12 +1,13 @@
 package com.example.group28assignmentapp;
 
-import androidx.appcompat.app.AppCompatActivity;
-import com.example.group28assignmentapp.databinding.ActivityMainBinding;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.group28assignmentapp.databinding.ActivityMainBinding;
+import com.example.group28assignmentapp.web_service.WebServiceActivity;
 
 public class MainActivity extends AppCompatActivity {
     private ActivityMainBinding binding;
@@ -18,8 +19,6 @@ public class MainActivity extends AppCompatActivity {
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         View view = binding.getRoot();
         setContentView(view);
-
-
 
         binding.atYourServiceButton.setOnClickListener(new View.OnClickListener() {
             @Override
