+++ app/src/main/java/com/example/group28assignmentapp/web_service/Category.java
@@ -1,4 +1,4 @@
-package com.example.group28assignmentapp;
+package com.example.group28assignmentapp.web_service;
 
 public enum Category {
     TOP_SONGS,
