+++ app/src/main/java/com/example/group28assignmentapp/web_service/Entry.java
@@ -1,4 +1,4 @@
-package com.example.group28assignmentapp;
+package com.example.group28assignmentapp.web_service;
 
 public class Entry {
     private String title;
