+++ app/src/main/java/com/example/group28assignmentapp/web_service/GenreGetFragment.java
@@ -1,25 +1,20 @@
-package com.example.group28assignmentapp;
+package com.example.group28assignmentapp.web_service;
 
-import android.content.Intent;
 import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
-import android.widget.ProgressBar;
 import android.widget.Spinner;
 
-import com.example.group28assignmentapp.databinding.FragmentGenreGetBinding;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
 
-import java.util.Timer;
-import java.util.TimerTask;
+import com.example.group28assignmentapp.R;
+import com.example.group28assignmentapp.databinding.FragmentGenreGetBinding;
 
 public class GenreGetFragment extends Fragment {
     private Spinner spinner;
