+++ app/src/main/java/com/example/group28assignmentapp/web_service/MainViewModel.java
@@ -1,6 +1,9 @@
-package com.example.group28assignmentapp;
+package com.example.group28assignmentapp.web_service;
 
 import androidx.lifecycle.ViewModel;
+
+import com.example.group28assignmentapp.web_service.Category;
+import com.example.group28assignmentapp.web_service.Entry;
 
 import java.util.ArrayList;
