+++ app/src/main/java/com/example/group28assignmentapp/web_service/SongListFragment.java
@@ -1,7 +1,10 @@
-package com.example.group28assignmentapp;
+package com.example.group28assignmentapp.web_service;
 
-import android.app.Dialog;
 import android.os.Bundle;
+import android.os.Handler;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
@@ -9,21 +12,11 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.os.Handler;
-import android.os.Looper;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ProgressBar;
-
-import com.airbnb.lottie.LottieDrawable;
 import com.example.group28assignmentapp.databinding.FragmentSongListBinding;
-import com.example.group28assignmentapp.databinding.ProgressDialogBinding;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -31,9 +24,6 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.concurrent.TimeUnit;
 
 import javax.net.ssl.HttpsURLConnection;
