+++ app/src/main/java/com/example/group28assignmentapp/web_service/WebServiceActivity.java
@@ -1,26 +1,17 @@
-package com.example.group28assignmentapp;
-
-import static android.view.View.Z;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.RecyclerView;
+package com.example.group28assignmentapp.web_service;
 
 import android.os.Bundle;
 import android.os.Handler;
 import android.view.View;
 import android.widget.TextView;
 
-import com.example.group28assignmentapp.databinding.ActivityMainBinding;
-import com.example.group28assignmentapp.databinding.ActivityWebServiceBinding;
+import androidx.appcompat.app.AppCompatActivity;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.ProtocolException;
+import com.example.group28assignmentapp.R;
+import com.example.group28assignmentapp.databinding.ActivityWebServiceBinding;
+import com.example.group28assignmentapp.web_service.GenreGetFragment;
+import com.example.group28assignmentapp.web_service.SongListFragment;
+
 import java.net.URL;
 
 public class WebServiceActivity extends AppCompatActivity {
