+++ app/src/main/res/layout/activity_web_service.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WebServiceActivity">
+    tools:context=".web_service.WebServiceActivity">
 
     <FrameLayout
         android:id="@+id/topPart"
