+++ app/src/main/res/layout/fragment_genre_get.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GenreGetFragment">
+    tools:context=".web_service.GenreGetFragment">
 
     <Spinner
         android:id="@+id/genreSpinner"
