+++ app/src/main/res/layout/fragment_song_list.xml
@@ -1,11 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SongListFragment">
+    tools:context=".web_service.SongListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/song_list_view"
