+++ app/src/test/java/com/example/group28assignmentapp/ExampleUnitTest.java
@@ -1,8 +1,8 @@
 package com.example.group28assignmentapp;
 
-import org.junit.Test;
+import static org.junit.Assert.assertEquals;
 
-import static org.junit.Assert.*;
+import org.junit.Test;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
